Designed specifically for Windows, so let's move to Windows
Plus additional fixes
